check for pre-set symlink

+++ install.sh
@@ -11,4 +11,9 @@ DOCKS_LIBDIR="/usr/share/docker-shorthands"
 
 mkdir -p "$DOCKS_LIBDIR"
 cp -r python/docks python/dockshort "$DOCKS_LIBDIR/"
-ln -s "$DOCKS_LIBDIR/docks" /usr/bin/docks
+
+if [ ! -e "$DOCKS_LIBDIR/docks" ]; then
+	ln -s "$DOCKS_LIBDIR/docks" /usr/bin/docks
+else
+	ls -l "/usr/bin/docks"
+fi
